requirements.txt
beautifulsoup4
lxml
html5lib
google-generativeai
Flask
requests
markdown